+++ errors.go
@@ -23,4 +23,6 @@ var (
 	ErrInvalidClientSecret    = errors.New("client secret is required")
 	ErrFailedToGetAccessToken = errors.New("failed to get access token")
 	ErrInvalidRefreshToken    = errors.New("invalid refresh token")
+	ErrInvalidCode            = errors.New("invalid code")
+	ErrGetAllWorkout          = errors.New("failed to get all workout")
 )