+++ responses.go
@@ -6,12 +6,58 @@ import (
 )
 
 type TokenResponse struct {
-	AccessToken  string `json:"access_token"`
-	TokenType    string `json:"token_type"`
-	ExpiresIn    int    `json:"expires_in"`
-	RefreshToken string `json:"refresh_token"`
+	AccessToken  string `json:"accessToken"`
+	TokenType    string `json:"tokenType"`
+	ExpiresIn    int    `json:"expiresIn"`
+	RefreshToken string `json:"refreshToken"`
 	Scope        string `json:"scope"`
-	CreatedAt    int64  `json:"created_at"`
+	CreatedAt    int64  `json:"createdAt"`
+}
+
+type WorkoutsResponse struct {
+	Workouts []Workout `json:"workouts"`
+	Total    int       `json:"total"`
+	Page     int       `json:"page"`
+	PerPage  int       `json:"per_page"`
+	Order    string    `json:"order"`
+	Sort     string    `json:"sort"`
+}
+
+type Workout struct {
+	Id             int            `json:"id"`
+	Starts         time.Time      `json:"starts"`
+	Minutes        int            `json:"minutes"`
+	Name           string         `json:"name"`
+	PlanId         *int           `json:"plan_id"`
+	WorkoutToken   string         `json:"workout_token"`
+	WorkoutTypeId  int            `json:"workout_type_id"`
+	WorkoutSummary WorkoutSummary `json:"workout_summary"`
+	CreatedAt      time.Time      `json:"created_at"`
+	UpdatedAt      time.Time      `json:"updated_at"`
+}
+
+type WorkoutSummary struct {
+	Id                  int         `json:"id"`
+	AscentAccum         string      `json:"ascent_accum"`
+	CadenceAvg          string      `json:"cadence_avg"`
+	CaloriesAccum       string      `json:"calories_accum"`
+	DistanceAccum       string      `json:"distance_accum"`
+	DurationActiveAccum string      `json:"duration_active_accum"`
+	DurationPausedAccum string      `json:"duration_paused_accum"`
+	DurationTotalAccum  string      `json:"duration_total_accum"`
+	HeartRateAvg        string      `json:"heart_rate_avg"`
+	PowerBikeNpLast     string      `json:"power_bike_np_last"`
+	PowerBikeTssLast    string      `json:"power_bike_tss_last"`
+	PowerAvg            string      `json:"power_avg"`
+	SpeedAvg            string      `json:"speed_avg"`
+	WorkAccum           string      `json:"work_accum"`
+	File                WorkoutFile `json:"file"`
+	CreatedAt           time.Time   `json:"created_at"`
+	UpdatedAt           time.Time   `json:"updated_at"`
+}
+
+type WorkoutFile struct {
+	Url string `json:"url"`
 }
 
 type ITokenResponse interface {
@@ -21,6 +67,17 @@ type ITokenResponse interface {
 	GetRefreshToken() string
 	GetScope() string
 	GetCreatedAt() time.Time
+}
+
+func UnmarshalToWorkoutsResponse(data []byte) (*WorkoutsResponse, *RequestError) {
+	var resp WorkoutsResponse
+	err := json.Unmarshal(data, &resp)
+
+	if err != nil {
+		return nil, NewError(err, 500, "failed to unmarshal response")
+	}
+
+	return &resp, nil
 }
 
 func UnmarshalToResponse(data []byte) (*TokenResponse, *RequestError) {
@@ -59,5 +116,5 @@ func (t TokenResponse) GetScope() string {
 }
 
 func (t TokenResponse) GetCreatedAt() time.Time {
-	return time.Unix(int64(t.CreatedAt), 0)
+	return time.Unix(t.CreatedAt, 0)
 }