+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Changelogs
 
-- Implement Wahoo Authentication
-  This release contains a new feature that allows users to authenticate with Wahoo. 
+- Implement Get All Workout API 
   
 test